chore: move monitoring
