+++ api/src/pdc/services/monitoring/contracts/statsrefresh.contract.ts
