+++ api/src/pdc/services/monitoring/contracts/statsrefresh.schema.ts
